requirements.txt
adk==0.0.5
altair==5.5.0
anyio==4.9.0
Authlib==1.6.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
cryptography==45.0.4
docstring_parser==0.16
et_xmlfile==2.0.0
fastapi==0.115.13
google-adk==1.4.1
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.98.0
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-genai==1.21.1
googleapis-common-protos==1.70.0
graphviz==0.21
grpcio==1.73.0
grpcio-status==1.73.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
MarkupSafe==3.0.2
nest-asyncio==1.6.0
numpy==2.3.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydeck==0.9.1
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rsa==4.9.1
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
streamlit==1.46.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.34.3
watchdog==6.0.0
websockets==15.0.1
